feat: clean up repo
