+++ i18n/en.pot
@@ -5,17 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2024-03-19T10:22:35.729Z\n"
-"PO-Revision-Date: 2024-03-19T10:22:35.729Z\n"
-
-msgid "Date"
-msgstr "Date"
-
-msgid "Height"
-msgstr "Height"
-
-msgid "Weight"
-msgstr "Weight"
+"POT-Creation-Date: 2024-03-21T15:11:37.305Z\n"
+"PO-Revision-Date: 2024-03-21T15:11:37.305Z\n"
 
 msgid "Growth Chart"
 msgstr "Growth Chart"
@@ -25,6 +16,15 @@ msgstr "There was an error fetching the config for the growth chart."
 
 msgid "Please check the configuration in Datastore Management and try again."
 msgstr "Please check the configuration in Datastore Management and try again."
+
+msgid "Date"
+msgstr "Date"
+
+msgid "Height"
+msgstr "Height"
+
+msgid "Weight"
+msgstr "Weight"
 
 msgid "Months"
 msgstr "Months"
