+++ src/Plugin.tsx
@@ -4,8 +4,8 @@ import './tailwind.css';
 import './index.css';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import i18n from '@dhis2/d2-i18n';
-import { WidgetCollapsible } from './Components/WidgetCollapsible';
-import { GrowthChart } from './Components/GrowthChart/GrowthChart';
+import { WidgetCollapsible } from './components/WidgetCollapsible';
+import { GrowthChart } from './components/GrowthChart/GrowthChart';
 import { EnrollmentOverviewProps } from './Plugin.types';
 import { useTeiById } from './utils/DataFetching/Hooks';
 import { useChartConfig } from './utils/DataFetching/Hooks/useChartConfig';
