+++ src/UI/GenericError/ChartConfigError.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import i18n from '@dhis2/d2-i18n';
-import { WidgetCollapsible } from '../../Components/WidgetCollapsible';
+import { WidgetCollapsible } from '../../components/WidgetCollapsible';
 import { Warning } from '../Icons';
 
 export const ChartConfigError = () => {
