--- src/utils/GrowthChartOptions/index.ts
@@ -1,2 +1,0 @@
-export { useGrowthChartAnnotations } from './useGrowthChartAnnotations';
-export type { AnnotationLabelType } from './useGrowthChartAnnotations';
