+++ src/utils/Hooks/Calculations/index.ts
@@ -1,0 +1,2 @@
+export { useCalculateDecimalDate } from './useCalculateDecimalDate';
+export { useCalculateMinMaxValues } from './useCalculateMinMaxValues';
