+++ src/utils/Hooks/Calculations/useCalculateMinMaxValues.ts
