+++ src/utils/Hooks/ChartDataVisualization/index.ts
@@ -1,3 +1,2 @@
 export { useZscoreLines } from './useZscoreLines';
 export { useMeasurementPlotting } from './useMeasurementPlotting';
-export { useGrowthChartAnnotations } from './GrowthChartOptions/useGrowthChartAnnotations';
