--- src/utils/useRangeTimePeriod.ts
@@ -1,4 +1,0 @@
-export const useRangeTimePeriod = (start: number, end: number) => Array.from(
-    { length: end - start + 1 },
-    (_, index) => start + index,
-);
